Split cmd between commands_raw and commands_extra
