+++ vulkano/src/command_buffer/cb/state_cache.rs
@@ -11,7 +11,7 @@ use std::sync::Arc;
 use command_buffer::cb::AddCommand;
 use command_buffer::cb::CommandBufferBuild;
 use command_buffer::CommandBufferBuilder;
-use command_buffer::cmd;
+use command_buffer::commands_raw;
 use command_buffer::DynamicState;
 use device::Device;
 use device::DeviceOwned;
@@ -80,13 +80,13 @@ unsafe impl<I> CommandBufferBuilder for StateCacheLayer<I>
 {
 }
 
-unsafe impl<Pl, I, O> AddCommand<cmd::CmdBindPipeline<Pl>> for StateCacheLayer<I>
-    where I: AddCommand<cmd::CmdBindPipeline<Pl>, Out = O>
+unsafe impl<Pl, I, O> AddCommand<commands_raw::CmdBindPipeline<Pl>> for StateCacheLayer<I>
+    where I: AddCommand<commands_raw::CmdBindPipeline<Pl>, Out = O>
 {
     type Out = StateCacheLayer<O>;
 
     #[inline]
-    fn add(mut self, command: cmd::CmdBindPipeline<Pl>) -> Self::Out {
+    fn add(mut self, command: commands_raw::CmdBindPipeline<Pl>) -> Self::Out {
         let raw_pipeline = command.sys().internal_object();
 
         let new_command = {
@@ -116,13 +116,13 @@ unsafe impl<Pl, I, O> AddCommand<cmd::CmdBindPipeline<Pl>> for StateCacheLayer<I
     }
 }
 
-unsafe impl<Cb, I, O> AddCommand<cmd::CmdExecuteCommands<Cb>> for StateCacheLayer<I>
-    where I: AddCommand<cmd::CmdExecuteCommands<Cb>, Out = O>
+unsafe impl<Cb, I, O> AddCommand<commands_raw::CmdExecuteCommands<Cb>> for StateCacheLayer<I>
+    where I: AddCommand<commands_raw::CmdExecuteCommands<Cb>, Out = O>
 {
     type Out = StateCacheLayer<O>;
 
     #[inline]
-    fn add(self, command: cmd::CmdExecuteCommands<Cb>) -> Self::Out {
+    fn add(self, command: commands_raw::CmdExecuteCommands<Cb>) -> Self::Out {
         // After a secondary command buffer is added, all states at reset to the "unknown" state.
         let new_inner = self.inner.add(command);
 
@@ -135,13 +135,13 @@ unsafe impl<Cb, I, O> AddCommand<cmd::CmdExecuteCommands<Cb>> for StateCacheLaye
     }
 }
 
-unsafe impl<I, O> AddCommand<cmd::CmdSetState> for StateCacheLayer<I>
-    where I: AddCommand<cmd::CmdSetState, Out = O>
+unsafe impl<I, O> AddCommand<commands_raw::CmdSetState> for StateCacheLayer<I>
+    where I: AddCommand<commands_raw::CmdSetState, Out = O>
 {
     type Out = StateCacheLayer<O>;
 
     #[inline]
-    fn add(mut self, command: cmd::CmdSetState) -> Self::Out {
+    fn add(mut self, command: commands_raw::CmdSetState) -> Self::Out {
         // We need to synchronize `self.dynamic_state` with the state in `command`.
         // While doing so, we tweak `command` to erase the states that are the same as what's
         // already in `self.dynamic_state`.
@@ -160,7 +160,7 @@ unsafe impl<I, O> AddCommand<cmd::CmdSetState> for StateCacheLayer<I>
         // TODO: missing implementations
 
         StateCacheLayer {
-            inner: self.inner.add(cmd::CmdSetState::new(command.device().clone(), command_state)),
+            inner: self.inner.add(commands_raw::CmdSetState::new(command.device().clone(), command_state)),
             dynamic_state: self.dynamic_state,
             graphics_pipeline: self.graphics_pipeline,
             compute_pipeline: self.compute_pipeline,
@@ -199,23 +199,23 @@ macro_rules! pass_through {
     }
 }
 
-pass_through!((Rp, F), cmd::CmdBeginRenderPass<Rp, F>);
-pass_through!((S, Pl), cmd::CmdBindDescriptorSets<S, Pl>);
-pass_through!((B), cmd::CmdBindIndexBuffer<B>);
-pass_through!((V), cmd::CmdBindVertexBuffers<V>);
-pass_through!((S, D), cmd::CmdBlitImage<S, D>);
-pass_through!((), cmd::CmdClearAttachments);
-pass_through!((S, D), cmd::CmdCopyBuffer<S, D>);
-pass_through!((S, D), cmd::CmdCopyBufferToImage<S, D>);
-pass_through!((S, D), cmd::CmdCopyImage<S, D>);
-pass_through!((), cmd::CmdDispatchRaw);
-pass_through!((), cmd::CmdDrawIndexedRaw);
-pass_through!((B), cmd::CmdDrawIndirectRaw<B>);
-pass_through!((), cmd::CmdDrawRaw);
-pass_through!((), cmd::CmdEndRenderPass);
-pass_through!((B), cmd::CmdFillBuffer<B>);
-pass_through!((), cmd::CmdNextSubpass);
-pass_through!((Pc, Pl), cmd::CmdPushConstants<Pc, Pl>);
-pass_through!((S, D), cmd::CmdResolveImage<S, D>);
-pass_through!((), cmd::CmdSetEvent);
-pass_through!((B, D), cmd::CmdUpdateBuffer<B, D>);
+pass_through!((Rp, F), commands_raw::CmdBeginRenderPass<Rp, F>);
+pass_through!((S, Pl), commands_raw::CmdBindDescriptorSets<S, Pl>);
+pass_through!((B), commands_raw::CmdBindIndexBuffer<B>);
+pass_through!((V), commands_raw::CmdBindVertexBuffers<V>);
+pass_through!((S, D), commands_raw::CmdBlitImage<S, D>);
+pass_through!((), commands_raw::CmdClearAttachments);
+pass_through!((S, D), commands_raw::CmdCopyBuffer<S, D>);
+pass_through!((S, D), commands_raw::CmdCopyBufferToImage<S, D>);
+pass_through!((S, D), commands_raw::CmdCopyImage<S, D>);
+pass_through!((), commands_raw::CmdDispatchRaw);
+pass_through!((), commands_raw::CmdDrawIndexedRaw);
+pass_through!((B), commands_raw::CmdDrawIndirectRaw<B>);
+pass_through!((), commands_raw::CmdDrawRaw);
+pass_through!((), commands_raw::CmdEndRenderPass);
+pass_through!((B), commands_raw::CmdFillBuffer<B>);
+pass_through!((), commands_raw::CmdNextSubpass);
+pass_through!((Pc, Pl), commands_raw::CmdPushConstants<Pc, Pl>);
+pass_through!((S, D), commands_raw::CmdResolveImage<S, D>);
+pass_through!((), commands_raw::CmdSetEvent);
+pass_through!((B, D), commands_raw::CmdUpdateBuffer<B, D>);
