+++ vulkano/src/command_buffer/commands_raw/copy_image.rs
