+++ vulkano/src/command_buffer/commands_raw/draw_indexed_raw.rs
