+++ vulkano/src/command_buffer/commands_raw/pipeline_barrier.rs
